Refactor out windows and darwin vfs

+++ src/fd_table.h
@@ -2,6 +2,7 @@
 #define FD_TABLE_H
 
 #include <stdbool.h>
+#include <stdint.h>
 
 #define TABLE_FD 0x3fa
 #define DEAD_FD 0x3f8
@@ -33,7 +34,7 @@ int become_remote_fd(int fd, int remote_fd);
 int become_local_fd(int fd, int local_fd);
 // lookup_real_fd looks up the real file descriptor and returns true if it's remote
 __attribute__((warn_unused_result))
-bool lookup_real_fd(int fd, int *out_real_fd);
+bool lookup_real_fd(int fd, intptr_t *out_real_fd);
 
 int perform_close(int fd);
 __attribute__((warn_unused_result))
